requirements.txt
asgiref==3.5.2
certifi==2022.9.14
charset-normalizer==2.1.1
cloudinary==1.29.0
Django==4.1
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
idna==3.4
Pillow==9.2.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.4.0
python-decouple==3.6
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
whitenoise==6.2.0
